moved import file code to own class
