+++ ownCloud/AppDelegate.swift
@@ -85,31 +85,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 	}
 
 	func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
-		let bookmarks : [OCBookmark] = OCBookmarkManager.shared.bookmarks as [OCBookmark]
-
-		if bookmarks.count > 1 {
-			let moreViewController = self.cardViewController(for: url)
-				if let wd = UIApplication.shared.delegate?.window {
-					let vc = wd!.rootViewController
-					if let navCon = vc as? UINavigationController, let topVC = navCon.visibleViewController {
-						OnMainThread {
-							topVC.present(asCard: moreViewController, animated: true)
-						}
-					} else {
-						OnMainThread {
-							vc?.present(asCard: moreViewController, animated: true)
-						}
-					}
-
-					return true
-				}
-		} else if bookmarks.count == 1, let bookmark = bookmarks.first {
-			self.importItemWithDirectoryPicker(with: url, into: bookmark)
-
-			return true
-		}
-
-		return false
+		return ImportFilesController(url: url).accountOrImportUI()
 	}
 
 	func application(_ application: UIApplication, performFetchWithCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void) {
@@ -132,91 +108,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 		Log.debug("AppDelegate: handle events for background URL session with identifier \(identifier)")
 
 		OCCoreManager.shared.handleEvents(forBackgroundURLSession: identifier, completionHandler: completionHandler)
-	}
-}
-
-extension AppDelegate {
-
-	func importItemWithDirectoryPicker(with url : URL, into bookmark: OCBookmark) {
-		OCCoreManager.shared.requestCore(for: bookmark, setup: { (_, _) in
-		}, completionHandler: { (core, error) in
-			if error == nil {
-				OnMainThread {
-					let directoryPickerViewController = ClientDirectoryPickerViewController(core: core!, path: "/", selectButtonTitle: "Save here".localized, avoidConflictsWith: [], choiceHandler: { [weak self] (selectedDirectory) in
-
-						if let targetDirectory = selectedDirectory {
-							self?.importFile(url: url, to: targetDirectory, core: core)
-						}
-					})
-
-					let pickerNavigationController = ThemeNavigationController(rootViewController: directoryPickerViewController)
-
-					if let wd = UIApplication.shared.delegate?.window {
-						let vc = wd!.rootViewController
-						if let navCon = vc as? UINavigationController, let topVC = navCon.visibleViewController {
-							topVC.present(pickerNavigationController, animated: true)
-						} else {
-							vc?.present(pickerNavigationController, animated: true)
-						}
-					}
-				}
-			}
-		})
-	}
-
-	func importFile(url : URL, to targetDirectory : OCItem, core : OCCore?) {
-		let name = url.lastPathComponent
-		if let progress = core?.importFileNamed(name,
-										 at: targetDirectory,
-										 from: url,
-										 isSecurityScoped: false,
-										 options: [OCCoreOption.importByCopying : true],
-										 placeholderCompletionHandler: { (error, item) in
-											if error != nil {
-												Log.debug("Error uploading \(Log.mask(name)) to \(Log.mask(targetDirectory.path)), error: \(error?.localizedDescription ?? "" )")
-											}
-											//placeholderHandler?(item, error)
-		},
-										 resultHandler: { (error, _ core, _ item, _) in
-											if error != nil {
-												Log.debug("Error uploading \(Log.mask(name)) to \(Log.mask(targetDirectory.path)), error: \(error?.localizedDescription ?? "" )")
-												//completionHandler?(false, item)
-											} else {
-												Log.debug("Success uploading \(Log.mask(name)) to \(Log.mask(targetDirectory.path))")
-												//completionHandler?(true, item)
-											}
-		}) {
-		} else {
-			Log.debug("Error setting up upload of \(Log.mask(name)) to \(Log.mask(targetDirectory.path))")
-		}
-	}
-
-	func cardViewController(for url: URL) -> UIViewController {
-		let tableViewController = MoreStaticTableViewController(style: .grouped)
-		let header = MoreViewHeader(url: url)
-		let moreViewController = MoreViewController(header: header, viewController: tableViewController)
-
-		let title = NSAttributedString(string: "Select Import Account".localized, attributes: [NSAttributedString.Key.font: UIFont.systemFont(ofSize: 20, weight: .heavy)])
-
-		var actionsRows: [StaticTableViewRow] = []
-
-		let bookmarks : [OCBookmark] = OCBookmarkManager.shared.bookmarks as [OCBookmark]
-
-		for (bookmark) in bookmarks {
-			let row = StaticTableViewRow(buttonWithAction: { [weak self] (_ row, _ sender) in
-				moreViewController.dismiss(animated: true, completion: nil)
-				self?.importItemWithDirectoryPicker(with: url, into: bookmark)
-			}, title: bookmark.shortName, style: .plain, image: Theme.shared.image(for: "owncloud-logo", size: CGSize(width: 25, height: 25)), imageWidth: 25, alignment: .left, identifier: "bookmark.uuid")
-			actionsRows.append(row)
-		}
-
-		let row = StaticTableViewRow(buttonWithAction: { (_ row, _ sender) in
-			moreViewController.dismiss(animated: true, completion: nil)
-		}, title: "Cancel".localized, style: .destructive, alignment: .center, identifier: "bookmark.uuid")
-		actionsRows.append(row)
-
-		tableViewController.addSection(MoreStaticTableViewSection(headerAttributedTitle: title, identifier: "actions-section", rows: actionsRows))
-
-		return moreViewController
 	}
 }
